Moving remainder of processing.py over to transform, deleting sum_gas_and_elec. Moving tests from test_processing.py to test_transform.py accordingly.
